Don't inline workhorse function, patch newer enumset
Using the newer enumset is like 25% faster for non-LTO builds.

Not inlining `astar_unchecked` is good because it's a function that does a
lot of work and generally neither it nor its callers can be simplified by
the inlining. The compiler made the wrong choice with this in the Dijkstra
benchmark, so adding the hint was a -10% improvement.